+++ trunk/clientgui/sg_StatImageLoader.cpp
@@ -70,7 +70,7 @@ void StatImageLoader::PopUpMenu(wxMouseEvent& WXUNUSED(event))
 #endif
 
 void StatImageLoader::RebuildMenu() {
-    for(int i=(int) statPopUpMenu->GetMenuItemCount()-1; i>=0;i--){
+    for (size_t i=statPopUpMenu->GetMenuItemCount()-1; i>=0;i--){
         wxMenuItem* item = statPopUpMenu->FindItemByPosition(i);
         statPopUpMenu->Delete(item);
     }
